Move linux System methods from linux/disk.go to linux/system.go.
